backend/requirements.txt
django>=6.0
djangorestframework>=3.17
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
Pillow>=10.0
